Move tests for sessions
Wihout any refactoring.
