+++ integration/session/killsessions_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package integration
+package session
 
 import (
 	"context"
@@ -29,6 +29,7 @@ import (
 	"github.com/FerretDB/FerretDB/v2/internal/util/must"
 	"github.com/FerretDB/FerretDB/v2/internal/util/testutil"
 
+	"github.com/FerretDB/FerretDB/v2/integration"
 	"github.com/FerretDB/FerretDB/v2/integration/setup"
 )
 
@@ -268,7 +269,7 @@ func killSessions(t testing.TB, ctx context.Context, conn *wireclient.Conn, dbNa
 	res, err := must.NotFail(resBody.(*wire.OpMsg).RawDocument()).DecodeDeep()
 	require.NoError(t, err)
 
-	fixCluster(t, res)
+	integration.FixCluster(t, res)
 
 	if expectedErr != nil {
 		testutil.AssertEqual(t, expectedErr, res)
